requirements.txt
dash==2.17.1
plotly==5.23.0
pandas==2.2.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1

